+++ src/deploy/functions/runtimes/index.ts
@@ -5,11 +5,11 @@ import * as validate from "../validate";
 import { FirebaseError } from "../../../error";
 
 /** Supported runtimes for new Cloud Functions. */
-const RUNTIMES: string[] = ["nodejs10", "nodejs12", "nodejs14", "nodejs16"];
+const RUNTIMES: string[] = ["nodejs10", "nodejs12", "nodejs14", "nodejs16", "nodejs18"];
 // Experimental runtimes are part of the Runtime type, but are in a
 // different list to help guard against some day accidentally iterating over
 // and printing a hidden runtime to the user.
-const EXPERIMENTAL_RUNTIMES = ["go113"];
+const EXPERIMENTAL_RUNTIMES: string[] = [];
 export type Runtime = typeof RUNTIMES[number] | typeof EXPERIMENTAL_RUNTIMES[number];
 
 /** Runtimes that can be found in existing backends but not used for new functions. */
@@ -33,7 +33,7 @@ const MESSAGE_FRIENDLY_RUNTIMES: Record<Runtime | DeprecatedRuntime, string> = {
   nodejs12: "Node.js 12",
   nodejs14: "Node.js 14",
   nodejs16: "Node.js 16",
-  go113: "Go 1.13",
+  nodejs18: "Node.js 18",
 };
 
 /**
@@ -61,6 +61,11 @@ export interface RuntimeDelegate {
    * the GCF API.
    */
   runtime: Runtime;
+
+  /**
+   * Path to the bin used to run the source code.
+   */
+  bin: string;
 
   /**
    * Validate makes sure the customers' code is actually viable.
@@ -108,8 +113,6 @@ export interface DelegateContext {
 }
 
 type Factory = (context: DelegateContext) => Promise<RuntimeDelegate | undefined>;
-// Note: golang has been removed from delegates because it does not work and it
-// is not worth having an experiment for yet.
 const factories: Factory[] = [node.tryCreateDelegate];
 
 /**