+++ src/test/emulators/auth/rest.spec.ts
@@ -93,6 +93,28 @@ describeAuthEmulator("authentication", ({ authApi }) => {
       });
   });
 
+  it("should accept API key as a query parameter", async () => {
+    await authApi()
+      .post("/identitytoolkit.googleapis.com/v1/accounts:signUp")
+      .query({ key: "fake-api-key" })
+      .send({})
+      .then((res) => {
+        expectStatusCode(200, res);
+        expect(res.body).not.to.have.property("error");
+      });
+  });
+
+  it("should accept API key in HTTP Header x-goog-api-key", async () => {
+    await authApi()
+      .post("/identitytoolkit.googleapis.com/v1/accounts:signUp")
+      .set("x-goog-api-key", "fake-api-key")
+      .send({})
+      .then((res) => {
+        expectStatusCode(200, res);
+        expect(res.body).not.to.have.property("error");
+      });
+  });
+
   it("should ignore non-Bearer Authorization headers", async () => {
     await authApi()
       .post("/identitytoolkit.googleapis.com/v1/accounts:signUp")
@@ -143,7 +165,11 @@ describeAuthEmulator("authentication", ({ authApi }) => {
       .post("/identitytoolkit.googleapis.com/v1/accounts:signUp")
       // This authenticates as owner of the default projectId. The exact value
       // and expiry don't matter -- the Emulator only checks for the format.
-      .set("Authorization", "Bearer ya29.AHES6ZRVmB7fkLtd1XTmq6mo0S1wqZZi3-Lh_s-6Uw7p8vtgSwg")
+      .set(
+        "Authorization",
+        // Not an actual token. Breaking it down to avoid linter false positives.
+        "Bearer ya" + "29.AHES0ZZZZZ0fff" + "ff0XXXX0mmmm0wwwww0-LL_l-0bb0b0bbbbbb"
+      )
       .send({
         // This field requires OAuth 2 and should work correctly.
         targetProjectId: "example2",