+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-- Fix bug where disabling background triggers did nothing. (#5221)
-- Fix bug in auth emulator where empty string should throw invalid email instead of missing email. (#3898)
-- Fix bug in auth emulator in which createdAt was not set for signInWithIdp new users. (#5203)
+- Refactor Functions Emulator. (#5422)
+- Fix race condition when discovering functions. (#5444)